pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-chaos-monkey"
version = "0.1.0"
description = "Autonomous AI agent for Kubernetes chaos engineering"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "langgraph>=0.2",
    "langchain-core>=0.3",
    "litellm>=1.40",
    "kubernetes>=30.1",
    "typer>=0.12",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "pyyaml>=6.0",
    "rich>=13.7",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[project.scripts]
chaos-monkey = "chaos_monkey.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/chaos_monkey"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
src = ["src"]
